So what kind of visual language can we use for the Spion / logo website. A first sketch for the poster reuses the kind of analytics you can easily keep yourself about your website.
I have a hard time thinking how to visualise the
concept of online privacy.

there is so much data available but this is mainly an
abstract concept.

Pierre if you say Google’s presentation of their data
does not resemble what kind of data they deal with,
what do you think it resembles? How could you show it?

The movies have their clichés.

0’s and 1’s and world maps.

But in the end it comes down to people watching
one another.

Panopticon? (A prison where all the prisoners can be
watched from one place, and they can’t see whether
they are being watched)

But on social networks the prisoners watch each
other, that’s the main fun of it

Is there a way to visualise the act of watching?

We have been thinking of making the website with information
that we scrape / search
Can we frame this information in such a way that it becomes
about watching?