Poster yes, check it out, read onwards for more (and the styles are back in local)
So you can run this poster with
python app.py

And visit the poster at http://127.0.0.1:5000/

You can drag the boxes around: everytime you release a box, a request is sent,
and the boxes’ position saved by the python script, to a json file, which in
turn is read when you reload the webpage: the locations are persistent.

The boxes’ inner html is also saved, along with its positions. So you can change
the contents in Firebug, and move a box to save. You can also edit the JSON directly.
The JSON you can also use to edit the size of the boxes.

There is a 10 by 10 grid; the vertical grid is the square root of 2 times as
high as the horizontal grid is wide: this is the ratio of the A series of papers.