Change navigation & no break on small screens & update spy-kit
Navigation no longer overlaps when scrolling: it stays in top.
(feedback Seda, for readability)

Also, on wide screens the layout is still fluid,
but a min-width of 1016 is maintained so it doesn’t break on small screens.

Finally, the spy-kit got a good tweak.
I just wonder if we don’t need to switch the background-color (yellow?).
Like this there is more differentation.